dataset/requirements.txt
absl-py==0.11.0
antlr4-python3-runtime==4.8
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
docopt==0.6.2
enum34==1.1.10
fairseq==0.9.0
filelock==3.0.12
google-api-core==1.26.1
google-auth==1.27.1
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-translate==3.1.0
google-cloud-vision==2.2.0
googleapis-common-protos==1.53.0
googletrans==4.0.0rc1
grpcio==1.36.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hydra-core==1.0.6
hyperframe==5.2.0
idna==2.10
img2pdf==0.4.0
importlib-metadata==3.7.0
importlib-resources==5.1.2
inflect==5.3.0
joblib==1.0.1
kiwisolver==1.3.1
libcst==0.3.17
lmdb==1.1.1
lorem-text==1.5
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.3.4
munkres==1.1.4
mypy-extensions==0.4.3
natsort==7.1.1
networkx==2.3
nlpaug==1.1.3
nltk==3.5
num2word==1.0.1
num2words==0.5.10
numpy==1.19.4
omegaconf==2.0.6
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
pandas-ml==0.6.1
pikepdf==2.11.4
Pillow==8.2.0
portalocker==2.0.0
proto-plus==1.15.0
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pyspellchecker==0.6.1
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
rfc3986==1.4.0
rsa==4.7.2
sacrebleu==1.5.1
sacremoses==0.0.45
scikit-learn==0.21.3
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
sniffio==1.2.0
snorkel==0.9.6
tensorboard==1.15.0
textblob==0.15.3
tokenizers==0.10.2
torch==1.8.0
torchvision==0.9.0
tqdm==4.58.0
transformers==4.5.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==1.0.1
zipp==3.4.1
